created a js folder in assets
