+++ app/javascript/application.js
@@ -3,4 +3,4 @@ import "@hotwired/turbo-rails"
 import "./controllers"
 import "bootstrap"
 import "@hotwired/stimulus"
-import "./crossword"
+// import "./crossword"
